Reworked the ID generation to use an unsigned long with an atomic increment function in the generator so it is thread safe.

+++ trunk/Libraries/OpenNeuronCL/IOpenNeuronCLBase.h
@@ -10,7 +10,7 @@ namespace OpenNeuronCL
 		IOpenNeuronCLBase(void)	{}
 		virtual ~IOpenNeuronCLBase(void) {};
 
-		virtual unsigned long ID() = 0;
+		virtual unsigned int ID() = 0;
 
 	};
